+++ arch/arm/cortex-m3/hr_kernel_tick_irq.c
@@ -1,0 +1,7 @@
+
+#include "hr_kernel_internal.h"
+
+void SysTick_Handler(void)
+{
+    hr_kernel_tick_from_isr();
+}