+++ arch/arm/cortex-m3/include/hr_port_config.h
@@ -1,0 +1,11 @@
+
+#ifndef HR_PORT_CONFIG_H
+#define HR_PORT_CONFIG_H
+
+#define HR_PORT_NAME                         "ARM Cortex-M3"
+#define HR_PORT_MIN_TASK_STACK_WORDS         18U
+#define HR_PORT_STACK_ALIGNMENT_BYTES        8U
+#define HR_PORT_SUPPORTS_FPU_CONTEXT          0
+#define HR_PORT_SUPPORTS_MPU                  0
+
+#endif /* HR_PORT_CONFIG_H */