+++ boards/bluepill_f103c8/board_clock.c
@@ -1,3 +1,4 @@
+
 #include "board.h"
 #include "stm32f1.h"
 
@@ -6,7 +7,7 @@ uint32_t board_get_core_clock_hz(void)
     return SystemCoreClock;
 }
 
-bool board_clock_is_72mhz(void)
+bool board_clock_is_nominal(void)
 {
     return SystemCoreClock == 72000000UL;
 }