+++ drivers/include/hr_hw_timer.h
@@ -1,9 +1,11 @@
+
 #ifndef HR_HW_TIMER_H
 #define HR_HW_TIMER_H
 
+#include <stdbool.h>
 #include <stdint.h>
 
-void hr_hw_timer_init_1khz(uint32_t core_clock_hz);
+bool hr_hw_timer_init(uint32_t tick_rate_hz);
 uint32_t hr_hw_timer_millis(void);
 void hr_hw_timer_delay_ms(uint32_t milliseconds);
 void hr_hw_timer_tick_isr(void);