+++ drivers/stm32f1/hr_gpio_stm32f1.c
@@ -1,64 +1,49 @@
+
 #include "hr_gpio.h"
+#include "hr_gpio_stm32f1.h"
 #include "stm32f1.h"
 
 #define GPIO_PIN_COUNT 16UL
+#define GPIO_PORT_SHIFT 8U
+#define GPIO_PORT_MASK  0xFFUL
+#define GPIO_PIN_MASK   0x1FUL
 
-static stm32f1_gpio_t *gpio_from_port(hr_gpio_port_t port)
+static uint32_t gpio_port_index(hr_gpio_pin_t pin)
 {
-    switch (port)
+    return ((uint32_t)pin >> GPIO_PORT_SHIFT) & GPIO_PORT_MASK;
+}
+
+static uint32_t gpio_pin_index(hr_gpio_pin_t pin)
+{
+    return (uint32_t)pin & GPIO_PIN_MASK;
+}
+
+static stm32f1_gpio_t *gpio_from_pin(hr_gpio_pin_t pin)
+{
+    switch (gpio_port_index(pin))
     {
-        case HR_GPIO_PORT_A:
+        case HR_STM32F1_GPIO_PORT_A:
             return STM32F1_GPIOA;
-        case HR_GPIO_PORT_B:
+        case HR_STM32F1_GPIO_PORT_B:
             return STM32F1_GPIOB;
-        case HR_GPIO_PORT_C:
+        case HR_STM32F1_GPIO_PORT_C:
             return STM32F1_GPIOC;
         default:
             return (stm32f1_gpio_t *)0;
     }
 }
 
-static void gpio_configure_nibble(stm32f1_gpio_t *gpio,
-                                  uint32_t pin,
-                                  uint32_t nibble)
+static void gpio_enable_port_clock(hr_gpio_pin_t pin)
 {
-    volatile uint32_t *config_register;
-    uint32_t shift;
-    uint32_t value;
-
-    if ((gpio == (stm32f1_gpio_t *)0) || (pin >= GPIO_PIN_COUNT))
+    switch (gpio_port_index(pin))
     {
-        return;
-    }
-
-    if (pin < 8UL)
-    {
-        config_register = &gpio->CRL;
-        shift = pin * 4UL;
-    }
-    else
-    {
-        config_register = &gpio->CRH;
-        shift = (pin - 8UL) * 4UL;
-    }
-
-    value = *config_register;
-    value &= ~(0x0FUL << shift);
-    value |= (nibble & 0x0FUL) << shift;
-    *config_register = value;
-}
-
-void hr_gpio_enable_port_clock(hr_gpio_port_t port)
-{
-    switch (port)
-    {
-        case HR_GPIO_PORT_A:
+        case HR_STM32F1_GPIO_PORT_A:
             STM32F1_RCC->APB2ENR |= RCC_APB2ENR_IOPAEN;
             break;
-        case HR_GPIO_PORT_B:
+        case HR_STM32F1_GPIO_PORT_B:
             STM32F1_RCC->APB2ENR |= RCC_APB2ENR_IOPBEN;
             break;
-        case HR_GPIO_PORT_C:
+        case HR_STM32F1_GPIO_PORT_C:
             STM32F1_RCC->APB2ENR |= RCC_APB2ENR_IOPCEN;
             break;
         default:
@@ -66,72 +51,157 @@ void hr_gpio_enable_port_clock(hr_gpio_port_t port)
     }
 }
 
-void hr_gpio_config_output_push_pull(hr_gpio_port_t port,
-                                     uint32_t pin,
-                                     hr_gpio_speed_t speed)
+static uint32_t gpio_mode_bits(hr_gpio_drive_t drive)
 {
-    gpio_configure_nibble(gpio_from_port(port), pin, (uint32_t)speed);
+    switch (drive)
+    {
+        case HR_GPIO_DRIVE_LOW:
+            return 2UL; /* STM32F1 output mode 2 MHz. */
+        case HR_GPIO_DRIVE_MEDIUM:
+            return 1UL; /* STM32F1 output mode 10 MHz. */
+        case HR_GPIO_DRIVE_HIGH:
+        case HR_GPIO_DRIVE_VERY_HIGH:
+            return 3UL; /* STM32F1 output mode 50 MHz. */
+        default:
+            return 2UL;
+    }
 }
 
-void hr_gpio_config_alternate_push_pull(hr_gpio_port_t port,
-                                        uint32_t pin,
-                                        hr_gpio_speed_t speed)
+static bool gpio_build_nibble(const hr_gpio_config_t *config,
+                              uint32_t *nibble)
 {
-    const uint32_t config = (2UL << 2U) | (uint32_t)speed;
-    gpio_configure_nibble(gpio_from_port(port), pin, config);
+    uint32_t value;
+
+    if ((config == (const hr_gpio_config_t *)0) ||
+        (nibble == (uint32_t *)0))
+    {
+        return false;
+    }
+
+    switch (config->mode)
+    {
+        case HR_GPIO_MODE_INPUT:
+            if (config->pull == HR_GPIO_PULL_NONE)
+            {
+                value = 0x4UL; /* Floating input. */
+            }
+            else
+            {
+                value = 0x8UL; /* Pull-up or pull-down input. */
+            }
+            break;
+        case HR_GPIO_MODE_OUTPUT_PUSH_PULL:
+            value = gpio_mode_bits(config->drive);
+            break;
+        case HR_GPIO_MODE_OUTPUT_OPEN_DRAIN:
+            value = gpio_mode_bits(config->drive) | 0x4UL;
+            break;
+        case HR_GPIO_MODE_ALTERNATE_PUSH_PULL:
+            value = gpio_mode_bits(config->drive) | 0x8UL;
+            break;
+        case HR_GPIO_MODE_ALTERNATE_OPEN_DRAIN:
+            value = gpio_mode_bits(config->drive) | 0xCUL;
+            break;
+        default:
+            return false;
+    }
+
+    *nibble = value;
+    return true;
 }
 
-void hr_gpio_config_input_floating(hr_gpio_port_t port, uint32_t pin)
+bool hr_gpio_configure(hr_gpio_pin_t pin, const hr_gpio_config_t *config)
 {
-    gpio_configure_nibble(gpio_from_port(port), pin, 4UL);
+    stm32f1_gpio_t *const gpio = gpio_from_pin(pin);
+    const uint32_t pin_index = gpio_pin_index(pin);
+    volatile uint32_t *config_register;
+    uint32_t shift;
+    uint32_t nibble;
+    uint32_t value;
+
+    if ((gpio == (stm32f1_gpio_t *)0) ||
+        (pin_index >= GPIO_PIN_COUNT) ||
+        !gpio_build_nibble(config, &nibble))
+    {
+        return false;
+    }
+
+    gpio_enable_port_clock(pin);
+
+    if (pin_index < 8UL)
+    {
+        config_register = &gpio->CRL;
+        shift = pin_index * 4UL;
+    }
+    else
+    {
+        config_register = &gpio->CRH;
+        shift = (pin_index - 8UL) * 4UL;
+    }
+
+    value = *config_register;
+    value &= ~(0x0FUL << shift);
+    value |= (nibble & 0x0FUL) << shift;
+    *config_register = value;
+
+    if ((config->mode == HR_GPIO_MODE_INPUT) &&
+        (config->pull != HR_GPIO_PULL_NONE))
+    {
+        hr_gpio_write(pin, config->pull == HR_GPIO_PULL_UP);
+    }
+
+    return true;
 }
 
-void hr_gpio_write(hr_gpio_port_t port, uint32_t pin, bool high)
+void hr_gpio_write(hr_gpio_pin_t pin, bool high)
 {
-    stm32f1_gpio_t *const gpio = gpio_from_port(port);
+    stm32f1_gpio_t *const gpio = gpio_from_pin(pin);
+    const uint32_t pin_index = gpio_pin_index(pin);
 
-    if ((gpio == (stm32f1_gpio_t *)0) || (pin >= GPIO_PIN_COUNT))
+    if ((gpio == (stm32f1_gpio_t *)0) || (pin_index >= GPIO_PIN_COUNT))
     {
         return;
     }
 
     if (high)
     {
-        gpio->BSRR = 1UL << pin;
+        gpio->BSRR = 1UL << pin_index;
     }
     else
     {
-        gpio->BRR = 1UL << pin;
+        gpio->BRR = 1UL << pin_index;
     }
 }
 
-bool hr_gpio_read(hr_gpio_port_t port, uint32_t pin)
+bool hr_gpio_read(hr_gpio_pin_t pin)
 {
-    stm32f1_gpio_t *const gpio = gpio_from_port(port);
+    stm32f1_gpio_t *const gpio = gpio_from_pin(pin);
+    const uint32_t pin_index = gpio_pin_index(pin);
 
-    if ((gpio == (stm32f1_gpio_t *)0) || (pin >= GPIO_PIN_COUNT))
+    if ((gpio == (stm32f1_gpio_t *)0) || (pin_index >= GPIO_PIN_COUNT))
     {
         return false;
     }
 
-    return (gpio->IDR & (1UL << pin)) != 0UL;
+    return (gpio->IDR & (1UL << pin_index)) != 0UL;
 }
 
-void hr_gpio_toggle(hr_gpio_port_t port, uint32_t pin)
+void hr_gpio_toggle(hr_gpio_pin_t pin)
 {
-    stm32f1_gpio_t *const gpio = gpio_from_port(port);
+    stm32f1_gpio_t *const gpio = gpio_from_pin(pin);
+    const uint32_t pin_index = gpio_pin_index(pin);
 
-    if ((gpio == (stm32f1_gpio_t *)0) || (pin >= GPIO_PIN_COUNT))
+    if ((gpio == (stm32f1_gpio_t *)0) || (pin_index >= GPIO_PIN_COUNT))
     {
         return;
     }
 
-    if ((gpio->ODR & (1UL << pin)) != 0UL)
+    if ((gpio->ODR & (1UL << pin_index)) != 0UL)
     {
-        gpio->BRR = 1UL << pin;
+        gpio->BRR = 1UL << pin_index;
     }
     else
     {
-        gpio->BSRR = 1UL << pin;
+        gpio->BSRR = 1UL << pin_index;
     }
 }