+++ drivers/stm32f1/include/hr_uart_stm32f1.h
@@ -1,0 +1,9 @@
+
+#ifndef HR_UART_STM32F1_H
+#define HR_UART_STM32F1_H
+
+#include "hr_uart.h"
+
+#define HR_STM32F1_UART1 ((hr_uart_instance_t)1U)
+
+#endif /* HR_UART_STM32F1_H */